Revertng 3db173de "publish bgs mock container (#3119)".
vro-mock-bgs-api is causing a few hundred SecRel vulnerabilities and as
we look to deploy our mock images into LHDI environments, there is now
the SecRel test image concept, which provides for significantly reduced
vulnerability requirements with the tradeoff that the images are only
approved for 2 weeks. Since we only need mock images for running
integration tests, two weeks is plenty of time for the longevity that we
need mock images to have.

This reverts commit 3db173de2a5593652852dbf1b0a72c2d665bf96c.